Refine dashboard login follow-up boundaries
