+++ packages/cli/src/hermes-setup.ts
@@ -1,7 +1,7 @@
 import type { Command } from 'commander';
 import { readFileSync } from 'node:fs';
 import { assertSelectableNetwork } from './config.js';
-import { ensureDashboardCredentialsForSetup } from './dashboard-credential-setup.js';
+import { ensureDashboardCredentialsForSetupBestEffort } from './dashboard-credential-setup.js';
 
 export type HermesMemoryMode = 'primary' | 'tools-only';
 
@@ -59,7 +59,7 @@ export interface HermesSetupActionDeps {
     opts: NormalizedHermesSetupOptions,
     runDeps?: {
       loadOpWallets?: (dir: string) => Promise<unknown>;
-      ensureDashboardCredentials?: (dkgHome: string) => Promise<unknown>;
+      afterConfigBootstrap?: (dkgHome: string) => Promise<unknown>;
     },
   ) => Promise<void>;
 }
@@ -127,8 +127,8 @@ export async function hermesSetupAction(
       nodeSkillContent: loadBundledDkgNodeSkill(),
     },
     {
-      ensureDashboardCredentials: (dkgHome: string) =>
-        ensureDashboardCredentialsForSetup(dkgHome, { prefix: '[hermes-setup]' }),
+      afterConfigBootstrap: (dkgHome: string) =>
+        ensureDashboardCredentialsForSetupBestEffort(dkgHome, { prefix: '[hermes-setup]' }),
       loadOpWallets: async (dir: string) => {
         const { loadOpWallets } = await import('@origintrail-official/dkg-agent');
         return loadOpWallets(dir);
