+++ packages/cli/src/openclaw-setup.ts
@@ -1,6 +1,6 @@
 import type { Command } from 'commander';
 import { assertSelectableNetwork } from './config.js';
-import { ensureDashboardCredentialsForSetup } from './dashboard-credential-setup.js';
+import { ensureDashboardCredentialsForSetupBestEffort } from './dashboard-credential-setup.js';
 
 /**
  * Options surface for the `dkg openclaw setup` subcommand as parsed by
@@ -26,7 +26,7 @@ export interface OpenClawSetupActionDeps {
   runSetup: (
     opts: OpenClawSetupCliOptions,
     runDeps?: {
-      ensureDashboardCredentials?: (dkgHome: string) => Promise<unknown>;
+      afterConfigBootstrap?: (dkgHome: string) => Promise<unknown>;
       loadOpWallets?: (dir: string) => Promise<unknown>;
     },
   ) => Promise<void>;
@@ -48,8 +48,8 @@ export async function openclawSetupAction(
   // adapter eagerly creates wallets before the daemon starts (issue #1306)
   // without the adapter package depending on dkg-agent.
   await deps.runSetup(opts, {
-    ensureDashboardCredentials: (dkgHome: string) =>
-      ensureDashboardCredentialsForSetup(dkgHome),
+    afterConfigBootstrap: (dkgHome: string) =>
+      ensureDashboardCredentialsForSetupBestEffort(dkgHome),
     loadOpWallets: async (dir: string) => {
       const { loadOpWallets } = await import('@origintrail-official/dkg-agent');
       return loadOpWallets(dir);
